configure: accept files for H5detect and H5make_libsettings
These tools previously wrote to stdout blindly. Instead, accept a
filepath as the first argument (using stdout if it does not exist) so
that anything doing improper things with stdout doesn't mess up line
endings.

THe problem appears when running the build with CTEST_USE_LAUNCHERS set
to TRUE which inserts a pipe between the stdout of H5detect and
H5make_libsettings and older CMake versions (< 3.2) where it reads in
binary, but writes in text causing \r\n to be turned into \r\r\n on
Windows.